some more cleanups
git-svn-id: https://svn.musicpd.org/mpd/trunk@60 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/ls.c
@@ -32,7 +32,7 @@
 #include <stdlib.h>
 
 char * dupAndStripPlaylistSuffix(char * file) {
-	int size = strlen(file)-strlen(PLAYLIST_FILE_SUFFIX)-1;
+	size_t size = strlen(file)-strlen(PLAYLIST_FILE_SUFFIX)-1;
 	char * ret = malloc(size+1);
 
 	strncpy(ret,file,size);
